stress: add `-bazel` support, support for sharding artifacts
When trying to integrate `stress` with Bazel, we run across a couple
issues.

* Tests are configured to write to an output file whose path is passed
  in via an environment variable, `XML_OUTPUT_FILE`.
  (Ref: https://docs.bazel.build/versions/main/test-encyclopedia.html)
  Since `stress` invokes the test process many times under the hood,
  those test processes are all going to write their `test.xml` output to
  the same location, which could cause data to be lost.
* Similarly, we [augment `rules_go`](https://github.com/cockroachdb/rules_go/commit/1ce9e349278b6e9ea1d52e0f73446c70ad440cbd)
  to have the test subprocesses store their `go test -json` output in a
  file whose path is passed in via the `GO_TEST_JSON_OUTPUT_FILE`
  environment variable. As above, the many test processes under `stress`
  collide with each other as they all try to write to the same file.

In this way, `stress`'s default behavior doesn't easily cooperate with
Bazel. `stress` assumes that all the meaningful output of a test is
captured via the `stdout`/`stderr` of the test processes, but in a
post-Bazel world where we are collecting test output that is configured
via environment variables, this does not work as well.

To this end we introduce a new flag `-shardable-artifacts`, which
specifies that certain environment variables reference test artifacts
that can be merged across test subprocesses. The usage of the flag is:

    -shardable-artifacts ENV_VAR1=prog1,ENV_VAR2=prog2

i.e., the argument is a comma-separated list of environment variable/
program pairs. When starting subprocesses, `stress` will set the
environment variables to a unique value for each process, then use the
given program to merge the produced artifacts. For example, with
`-shardable-artifacts FOO=cat`, `stress` will give each subprocess its
own value of `FOO`, and then it will "merge" all the `FOO` files for the
failing tests by `cat`ing them.

Additionally we provide a command-line argument `-bazel` to propagate
the environment variable Bazel uses, `TEST_TMPDIR`, down to the standard
`TMPDIR` variable for convenience.